requirements.txt
aiodns==1.1.1
aiohttp==3.5.1
aniso8601==7.0.0
astroid==2.1.0
async-timeout==3.0.1
attrs==18.2.0
autopep8==1.4.3
ccxt==1.18.87
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
cytoolz==0.10.0
defusedxml==0.6.0
Django==2.1.5
django-allauth==0.40.0
django-el-pagination==3.2.4
django-filter==2.2.0
django-graphiql==0.4.4
django-graphql-jwt==0.2.2
django-graphql-social-auth==0.1.4
environ==1.0
flake8==3.6.0
Flask==1.0.2
Flask-MySQLdb==0.2.0
future==0.17.1
graphene==2.1.8
graphene-django==2.6.0
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.7
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
multidict==4.5.2
multipledispatch==0.5.0
mysqlclient==1.3.14
numpy==1.15.4
oauthlib==3.1.0
pandas==0.23.4
parso==0.3.1
Pillow==5.4.1
pipenv==2018.11.26
pluggy==0.8.0
promise==2.2.1
psycopg2-binary==2.7.6.1
pycares==2.3.0
pycodestyle==2.4.0
pycparser==2.19
pycryptodome==3.8.2
pydocstyle==3.0.0
pyflakes==2.0.0
PyJWT==1.7.1
pylint==2.2.2
pylint-django==2.0.11
pylint-plugin-utils==0.6
pymongo==3.9.0
python-dateutil==2.7.5
python-jsonrpc-server==0.0.2
python-language-server==0.21.4
python3-openid==3.1.0
pytz==2018.9
requests==2.20.0
requests-cache==0.4.13
requests-oauthlib==1.2.0
rope==0.11.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
toolz==0.10.0
urllib3==1.24.3
virtualenv==16.1.0
virtualenv-clone==0.4.0
Werkzeug==0.14.1
wrapt==1.10.11
yapf==0.25.0
yarl==1.1.0
